`Add "Transaction Header" section to TransactionHeader.md`
This commit includes the addition of a new section titled "Transaction Header" in the `TransactionHeader.md` file. The section provides a detailed description of the transaction header object, which contains metadata for a transaction. A table has been added to detail the properties of the transaction header, including `transactionId`, `transactionType`, `sourceId`, and `sourceName`. All these properties are marked as required. The commit also includes explanations for each property, such as the `transactionId` being a GUID and the `transactionType` being an enumeration of possible transaction types.